+++ android/src/main/kotlin/com/bregant/azure_speech_recognition/AzureSpeechRecognitionPlugin.kt
@@ -17,6 +17,10 @@ import com.microsoft.cognitiveservices.speech.SpeechRecognitionResult;
 import com.microsoft.cognitiveservices.speech.SpeechRecognizer;
 import com.microsoft.cognitiveservices.speech.CancellationDetails;
 import com.microsoft.cognitiveservices.speech.KeywordRecognitionModel;
+import com.microsoft.cognitiveservices.speech.PropertyId;
+import com.microsoft.cognitiveservices.speech.transcription.ConversationTranscriber;
+import com.microsoft.cognitiveservices.speech.transcription.ConversationTranscriptionEventArgs;
+import com.microsoft.cognitiveservices.speech.transcription.ConversationTranscriptionResult;
 import com.bregant.azure_speech_recognition.MicrophoneStream;
 import android.app.Activity;
 
@@ -103,61 +107,75 @@ public class AzureSpeechRecognitionPlugin(): FlutterPlugin,Activity(),MethodCall
       var permissionRequestId : Int = 5;
       var speechSubscriptionKey : String = ""+call.argument("subscriptionKey");
       var serviceRegion : String= ""+call.argument("region");
+      var endpoint : String? = call.argument("endpoint");
       var lang : String = ""+call.argument("language");
-
-      simpleSpeechRecognition(speechSubscriptionKey,serviceRegion,lang);
+      simpleSpeechRecognition(speechSubscriptionKey,serviceRegion,endpoint,lang);
       result.success(true);
 
     }else if(call.method == "micStream"){
       var permissionRequestId : Int = 5;
       var speechSubscriptionKey : String = ""+call.argument("subscriptionKey");
       var serviceRegion : String= ""+call.argument("region");
+      var endpoint : String? = call.argument("endpoint");
       var lang : String = ""+call.argument("language");
 
 
-      micStreamRecognition(speechSubscriptionKey,serviceRegion,lang);
+      micStreamRecognition(speechSubscriptionKey,serviceRegion,endpoint,lang);
       result.success(true);
 
     }else if(call.method == "continuousStream"){
       var permissionRequestId : Int = 5;
       var speechSubscriptionKey : String = ""+call.argument("subscriptionKey");
       var serviceRegion : String= ""+call.argument("region");
+      var endpoint : String? = call.argument("endpoint");
       var lang : String =  ""+call.argument("language");
 
 
-      micStreamContinuosly(speechSubscriptionKey,serviceRegion,lang);
+      micStreamContinuosly(speechSubscriptionKey,serviceRegion,endpoint,lang);
       result.success(true);
 
     }else if(call.method == "dictationMode"){
       var permissionRequestId : Int = 5;
       var speechSubscriptionKey : String = ""+call.argument("subscriptionKey");
       var serviceRegion : String= ""+call.argument("region");
+      var endpoint : String? = call.argument("endpoint");
       var lang : String =  ""+call.argument("language");
 
       enableDictation = true;
-      micStreamContinuosly(speechSubscriptionKey,serviceRegion,lang);
+      micStreamContinuosly(speechSubscriptionKey,serviceRegion,endpoint,lang);
       result.success(true);
 
-    }
-    else if(call.method == "intentRecognizer"){
+    }else if(call.method == "transcribeWithDiarization"){
       var permissionRequestId : Int = 5;
       var speechSubscriptionKey : String = ""+call.argument("subscriptionKey");
       var serviceRegion : String= ""+call.argument("region");
+      var endpoint : String? = call.argument("endpoint");
+      var lang : String =  ""+call.argument("language");
+
+      transcribeWithDiarization(speechSubscriptionKey, serviceRegion, endpoint, lang);
+      result.success(true);
+
+    } else if(call.method == "intentRecognizer"){
+      var permissionRequestId : Int = 5;
+      var speechSubscriptionKey : String = ""+call.argument("subscriptionKey");
+      var serviceRegion : String= ""+call.argument("region");
+      var endpoint : String? = call.argument("endpoint");
       var appId : String= ""+call.argument("appId");
       var lang : String =  ""+call.argument("language");
 
 
-      recognizeIntent(speechSubscriptionKey,serviceRegion,appId,lang);
+      recognizeIntent(speechSubscriptionKey,serviceRegion,endpoint,appId,lang);
       result.success(true);
 
     }else if(call.method == "keywordRecognizer"){
       var permissionRequestId : Int = 5;
       var speechSubscriptionKey : String = ""+call.argument("subscriptionKey");
       var serviceRegion : String= ""+call.argument("region");
+      var endpoint : String? = call.argument("endpoint");
       var lang : String =  ""+call.argument("language");
       var kwsModel : String = ""+call.argument("kwsModel");
 
-      keywordRecognizer(speechSubscriptionKey, serviceRegion, lang, kwsModel);
+      keywordRecognizer(speechSubscriptionKey, serviceRegion, endpoint, lang, kwsModel);
       result.success(true);
 
     }
@@ -170,7 +188,11 @@ public class AzureSpeechRecognitionPlugin(): FlutterPlugin,Activity(),MethodCall
     azureChannel.setMethodCallHandler(null)
   }
 
-  fun simpleSpeechRecognition(speechSubscriptionKey:String,serviceRegion:String,lang:String) {
+  fun simpleSpeechRecognition(
+    speechSubscriptionKey:String,
+    serviceRegion:String,
+    endpoint:String?,
+    lang:String) {
     val logTag : String = "simpleVoice";
 
 
@@ -178,7 +200,11 @@ public class AzureSpeechRecognitionPlugin(): FlutterPlugin,Activity(),MethodCall
       
       var audioInput : AudioConfig = AudioConfig.fromStreamInput(createMicrophoneStream());
 
-      var config : SpeechConfig = SpeechConfig.fromSubscription(speechSubscriptionKey, serviceRegion); 
+      var config : SpeechConfig = if(endpoint != null && endpoint.isNotEmpty() && endpoint != "null"){
+        SpeechConfig.fromEndpoint(URI(endpoint), speechSubscriptionKey);
+      }else{
+        SpeechConfig.fromSubscription(speechSubscriptionKey, serviceRegion);
+      }
       assert(config != null);
 
       config.speechRecognitionLanguage = lang;
@@ -216,7 +242,11 @@ public class AzureSpeechRecognitionPlugin(): FlutterPlugin,Activity(),MethodCall
 
 
   // Mic Streaming, it need the additional method implementend to get the data from the async task
-  fun micStreamRecognition(speechSubscriptionKey:String,serviceRegion:String,lang:String){
+  fun micStreamRecognition(
+    speechSubscriptionKey:String,
+    serviceRegion:String,
+    endpoint:String?,
+    lang:String){
     val logTag : String = "micStream";
 
     try{
@@ -224,7 +254,11 @@ public class AzureSpeechRecognitionPlugin(): FlutterPlugin,Activity(),MethodCall
       var audioInput : AudioConfig = AudioConfig.fromStreamInput(createMicrophoneStream());
 
 
-      var config : SpeechConfig = SpeechConfig.fromSubscription(speechSubscriptionKey, serviceRegion); 
+      var config : SpeechConfig = if(endpoint != null && endpoint.isNotEmpty() && endpoint != "null"){
+        SpeechConfig.fromEndpoint(URI(endpoint), speechSubscriptionKey);
+      }else{
+        SpeechConfig.fromSubscription(speechSubscriptionKey, serviceRegion);
+      }
       assert(config != null);
 
       config.speechRecognitionLanguage = lang;
@@ -262,7 +296,11 @@ public class AzureSpeechRecognitionPlugin(): FlutterPlugin,Activity(),MethodCall
 
   // stream continuosly until you press the button to stop ! STILL NOT WORKING COMPLETELY
 
-  fun micStreamContinuosly(speechSubscriptionKey:String,serviceRegion:String,lang:String){
+  fun micStreamContinuosly(
+    speechSubscriptionKey:String,
+    serviceRegion:String,
+    endpoint:String?,
+    lang:String){
     val logTag : String = "micStreamContinuos";
     
     
@@ -298,7 +336,11 @@ public class AzureSpeechRecognitionPlugin(): FlutterPlugin,Activity(),MethodCall
       //audioInput = AudioConfig.fromStreamInput(createMicrophoneStream());
 
 
-      var config : SpeechConfig = SpeechConfig.fromSubscription(speechSubscriptionKey, serviceRegion); 
+      var config : SpeechConfig = if(endpoint != null && endpoint.isNotEmpty() && endpoint != "null"){
+        SpeechConfig.fromEndpoint(URI(endpoint), speechSubscriptionKey);
+      }else{
+        SpeechConfig.fromSubscription(speechSubscriptionKey, serviceRegion);
+      }
       assert(config != null);
 
       config.speechRecognitionLanguage = lang;
@@ -356,7 +398,12 @@ public class AzureSpeechRecognitionPlugin(): FlutterPlugin,Activity(),MethodCall
 
   /// Recognize Intent method from microsoft sdk
 
-  fun recognizeIntent(speechSubscriptionKey:String,serviceRegion:String,appId:String,lang:String){
+  fun recognizeIntent(
+    speechSubscriptionKey:String,
+    serviceRegion:String,
+    endpoint:String?,
+    appId:String,
+    lang:String){
     val logTag : String = "intent";
 
     var content :  ArrayList<String> = ArrayList<String>();
@@ -369,7 +416,11 @@ public class AzureSpeechRecognitionPlugin(): FlutterPlugin,Activity(),MethodCall
       val audioInput = AudioConfig.fromStreamInput(createMicrophoneStream());
 
 
-      var config : SpeechConfig = SpeechConfig.fromSubscription(speechSubscriptionKey, serviceRegion); 
+      var config : SpeechConfig = if(endpoint != null && endpoint.isNotEmpty() && endpoint != "null"){
+        SpeechConfig.fromEndpoint(URI(endpoint), speechSubscriptionKey);
+      }else{
+        SpeechConfig.fromSubscription(speechSubscriptionKey, serviceRegion);
+      }
 
       assert(config != null);
 
@@ -419,7 +470,12 @@ public class AzureSpeechRecognitionPlugin(): FlutterPlugin,Activity(),MethodCall
     }
   }
 
-  fun keywordRecognizer(speechSubscriptionKey:String,serviceRegion:String,lang:String,kwsModelFile:String) {
+  fun keywordRecognizer(
+    speechSubscriptionKey:String,
+    serviceRegion:String,
+    endpoint:String?,
+    lang:String,
+    kwsModelFile:String) {
     val logTag : String = "keyword";
     var continuousListeningStarted : Boolean = false;
     lateinit var reco : SpeechRecognizer;
@@ -451,7 +507,11 @@ public class AzureSpeechRecognitionPlugin(): FlutterPlugin,Activity(),MethodCall
 
       audioInput = AudioConfig.fromStreamInput(createMicrophoneStream());
 
-      var config : SpeechConfig = SpeechConfig.fromSubscription(speechSubscriptionKey, serviceRegion); 
+      var config : SpeechConfig = if(endpoint != null && endpoint.isNotEmpty() && endpoint != "null"){
+        SpeechConfig.fromEndpoint(URI(endpoint), speechSubscriptionKey);
+      }else{
+        SpeechConfig.fromSubscription(speechSubscriptionKey, serviceRegion);
+      }
 
       assert(config != null);
 
@@ -529,6 +589,45 @@ public class AzureSpeechRecognitionPlugin(): FlutterPlugin,Activity(),MethodCall
 
     handler.post{
         azureChannel.invokeMethod(method,arguments); 
+    }
+  }
+
+  fun transcribeWithDiarization(
+    speechSubscriptionKey:String,
+    serviceRegion:String,
+    endpoint:String?,
+    lang:String) {
+
+    val logTag : String = "transcriber"
+    try{
+      val audioInput = AudioConfig.fromStreamInput(createMicrophoneStream());
+      val config : SpeechConfig = if(endpoint != null && endpoint.isNotEmpty() && endpoint != "null"){
+        SpeechConfig.fromEndpoint(URI(endpoint), speechSubscriptionKey);
+      }else{
+        SpeechConfig.fromSubscription(speechSubscriptionKey, serviceRegion);
+      }
+
+      config.speechRecognitionLanguage = lang;
+      config.setProperty(PropertyId.SpeechServiceResponse_DiarizeIntermediateResults, "true");
+
+      val transcriber = ConversationTranscriber(config, audioInput);
+
+      invokeMethod("speech.onRecognitionStarted",null);
+
+      transcriber.transcribing.addEventListener({ o, args ->
+        val txt = args.getResult().getText();
+        val speaker = args.getResult().getSpeakerId();
+        invokeMethod("speech.onSpeech","$speaker:$txt");
+      });
+
+      val task : Future<Void> = transcriber.startTranscribingAsync();
+
+      setOnTaskCompletedListener(task, { result ->
+        // started
+      })
+
+    }catch(exec:Exception){
+      invokeMethod("speech.onException", "Exception: "+exec.message);
     }
   }